+++ rust/crates/openh264-rs/tests/error_concealment_api_test.rs
@@ -20,7 +20,7 @@ fn test_decoder_error_concealment_modes() {
         unsafe {
             let mut p_decoder: *mut ISVCDecoder = std::ptr::null_mut();
             let ret = WelsCreateDecoder(&mut p_decoder);
-            assert_eq!(ret, CM_RESULT_SUCCESS as i64);
+            assert_eq!(i64::from(ret), CM_RESULT_SUCCESS as i64);
             assert!(!p_decoder.is_null());
 
             let mut param = SDecodingParam::default();
@@ -28,17 +28,17 @@ fn test_decoder_error_concealment_modes() {
             param.eEcActiveIdc = ec_mode;
 
             let init_ret = (*p_decoder).Initialize(&param as *const SDecodingParam);
-            assert_eq!(init_ret, CM_RESULT_SUCCESS as i64);
+            assert_eq!(i64::from(init_ret), CM_RESULT_SUCCESS as i64);
 
             let mut current_ec: i32 = 0;
             let get_opt_ret = (*p_decoder).GetOption(
                 DECODER_OPTION::DECODER_OPTION_ERROR_CON_IDC,
                 &mut current_ec as *mut i32 as *mut std::ffi::c_void,
             );
-            assert_eq!(get_opt_ret, CM_RESULT_SUCCESS as i64);
+            assert_eq!(i64::from(get_opt_ret), CM_RESULT_SUCCESS as i64);
 
             let uninit_ret = (*p_decoder).Uninitialize();
-            assert_eq!(uninit_ret, CM_RESULT_SUCCESS as i64);
+            assert_eq!(i64::from(uninit_ret), CM_RESULT_SUCCESS as i64);
 
             WelsDestroyDecoder(p_decoder);
         }
